Split MainWindow modules
